Fix a bug in the CPU frequency function.  The type is a uint32 not a unit16.
Submitted by:   Meno Abels <meno.abels@adviser.com>

+++ sysutils/ganglia-monitor-core/files/patch-srclib_libmetrics_freebsd_metrics.c
@@ -58,7 +58,7 @@ $FreeBSD$
 +   }
 +
 +done:
-+   val.uint16 = freq;
++   val.uint32 = freq;
  
     return val;
  }
